test/ruby/test_process.rb: test for fd=3 usability in child
Ensure we can redirect anything to fd=3 in a child process.  This
test exists because fd=3 is a commonly reserved FD for a timer
thread pipe in the parent Ruby VM, but fd=3 is the first FD used by
the sd_listen_fds function for systemd.

This means there is a possibility for a bug to slip in where the
redirect to fd=3 fails to work.  This test should ensure the
continued viability of systemd replacements in Ruby :>

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@51146 b2dd03c8-39d4-4d8f-98ff-823fe69b080e