Pin objects that have the write barrier executed on them
During incremental marking, if a black object gets a new reference to a
white object, we'll execute the write barrier and grey the white object.
The greying process pushes the white object on the mark stack, keeping
it alive.  The problem is that the write barrier doesn't know what mark
function the black object _would have_ used if it were to mark its
references using its mark function.  Since we cannot know what mark
function the object would have used, for safety, we need to assume that
the object should be pinned.